requirements.txt
streamlit>=1.30.0
openai>=1.10.0
Pillow>=10.0.0
pandas>=2.0.0
qrcode[pil]>=7.4.0
